Fixed the issues with the visited links and nav bar in IE
Both issues seem (to me) to be bugs in the browsers. The invisible visited links are white on white, but the white color should be only in the nav links, not in body. And the picture was sized correctly, and then enlarged when made a hyperlink, but only in IE. Ready to go live now.